add trycatch error on deletePlayer

+++ src/server/controllers/playerControllers.js
@@ -1,4 +1,3 @@
-const { default: ObjectID } = require("bson-objectid");
 const jwt = require("jsonwebtoken");
 const Player = require("../../database/models/player");
 const User = require("../../database/models/user");
@@ -28,20 +27,35 @@ const deletePlayer = async (req, res, next) => {
   const headerAuthorization = req.header("authorization");
   const token = headerAuthorization.replace("Bearer ", "");
   const { username } = jwt.decode(token);
-  const { id } = req.params;
-  await Player.findByIdAndDelete(id);
-  const user = await User.findOne({ username });
-  const userUpdated = await User.findOneAndUpdate(
-    { username },
-    {
-      players: (user.players = user.players.filter(
-        (player) => player.toString() !== id
-      )),
-    },
-    { new: true }
-  );
 
-  res.json(userUpdated);
+  try {
+    const { id } = req.params;
+
+    const user = await User.findOne({ username });
+    const userUpdated = await User.findOneAndUpdate(
+      { username },
+      {
+        players: user.players.filter((player) => player.toString() !== id),
+      },
+      { new: true }
+    );
+    if (!userUpdated) {
+      const error = new Error("This player doesn't belong to your user");
+      error.code = 403;
+      return next(error);
+    }
+
+    const deletedPlayer = await Player.findByIdAndDelete(id);
+    if (!deletedPlayer) {
+      const error = new Error("This player doesn't exist");
+      error.code = 404;
+      return next(error);
+    }
+
+    return res.json(userUpdated);
+  } catch (error) {
+    return next(error);
+  }
 };
 
 module.exports = { createPlayer, deletePlayer };
